+++ karaf-system/system.ocr/src/main/java/patlego/vm/github/io/ocr/utils/TesseractConversionResult.java
@@ -8,7 +8,7 @@
  * Copyright (c) 2020 LegoTech
  */
 
-package patlego.vm.github.io.tesseract.utils;
+package patlego.vm.github.io.ocr.utils;
 
 import java.io.InputStream;
